Ugly hack for not overwriting remembered values in Firefox.
Note: this commit is not meant for submission to Angular, as it's
missing tests, is probably the wrong approach, etc.--it just works
for us. If anyone wants to clean it up for submission, please do.

For background see:

https://github.com/angular/angular.js/issues/1460

Also note: this only fixes Firefox, since it auto-fills remembered values
before onload runs. For Chrome, see:

https://code.google.com/p/chromium/issues/detail?id=224405